+++ ddb/feature/smartcd/__init__.py
@@ -1,9 +1,10 @@
 # -*- coding: utf-8 -*-
-from typing import Iterable
+from typing import Iterable, ClassVar
 
 from ddb.action import Action
 from ddb.feature import Feature
 from .actions import SmartcdAction, WindowsProjectActivate
+from .schema import SmartcdSchema
 
 
 class SmartcdFeature(Feature):
@@ -14,6 +15,10 @@ class SmartcdFeature(Feature):
     @property
     def name(self) -> str:
         return "smartcd"
+
+    @property
+    def schema(self) -> ClassVar[SmartcdSchema]:
+        return SmartcdSchema
 
     @property
     def actions(self) -> Iterable[Action]: